+++ mantra/transforms/__init__.py
@@ -6,14 +6,12 @@ from .encodings import (
     NodeRandomTransform,
     SimplexRandomTransform,
 )
-from .node_task_transforms import (
-    AttributeToNodeClassTransform,
-    AttributeToNodeRegressionTransform,
-)
 from .task_transforms import (
     NAME_TO_CLASS_2M,
     NAME_TO_CLASS_3M,
     AttributeToClassTransform,
+    AttributeToNodeClassTransform,
+    AttributeToNodeRegressionTransform,
     AttributeToRegressionTransform,
     BettiToClassTransform,
     NameToClass2MTransform,